requirements.txt
langgraph==0.2.16
langchain==0.3.0
langchain-core==0.3.0
typing-extensions==4.12.0
fastmcp==0.2.0
aiohttp==3.8.6
pydantic>=2.5.3
PyYAML==6.0
